♻️ Refactor: CreateGroup 관련 컴포넌트 위치를 components/create-group 폴더로 이동 (#72)
