--- src/components/common/group/index.ts
@@ -1,4 +1,0 @@
-export { default as BasicInfoSection } from '@/components/common/group/BasicInfoSection'
-export { default as PeriodMembersSection } from '@/components/common/group/PeriodMembersSection'
-export { default as LecturePickerSection } from '@/components/common/group/LecturePickerSection'
-// export { default as DateModal } from '@/components/common/group/DatePickerModal'
