+++ src/components/common/index.ts
@@ -11,4 +11,4 @@ export * from '@components/common/markdown-editor'
 export * from '@components/common/skeleton'
 export * from '@components/common/card'
 export * from '@/components/common/date-picker'
-export * from '@/components/common/group'
+export * from '@/components/create-group'
