+++ src/components/create-group/BasicInfoSection.tsx
