+++ src/components/create-group/LecturePickerSection.tsx
