+++ src/components/create-group/index.ts
@@ -1,0 +1,4 @@
+export { default as BasicInfoSection } from '@/components/create-group/BasicInfoSection'
+export { default as PeriodMembersSection } from '@/components/create-group/PeriodMembersSection'
+export { default as LecturePickerSection } from '@/components/create-group/LecturePickerSection'
+// export { default as DateModal } from '@/components/common/group/DatePickerModal'
